refactor(config): move server related config into own file
