+++ internal/config/server.go
@@ -1,0 +1,24 @@
+package config
+
+const (
+	// configuration keys
+	serverHost      = "server.host"
+	serverProtocol  = "server.protocol"
+	serverHTTPPort  = "server.http_port"
+	serverHTTPSPort = "server.https_port"
+	serverGRPCPort  = "server.grpc_port"
+	serverCertFile  = "server.cert_file"
+	serverCertKey   = "server.cert_key"
+)
+
+// ServerConfig contains fields, which configure both HTTP and gRPC
+// API serving.
+type ServerConfig struct {
+	Host      string `json:"host,omitempty"`
+	Protocol  Scheme `json:"protocol,omitempty"`
+	HTTPPort  int    `json:"httpPort,omitempty"`
+	HTTPSPort int    `json:"httpsPort,omitempty"`
+	GRPCPort  int    `json:"grpcPort,omitempty"`
+	CertFile  string `json:"certFile,omitempty"`
+	CertKey   string `json:"certKey,omitempty"`
+}
